GH-231 Added artifactory_virtual_conan_repository resource
