+++ pkg/artifactory/resource_artifactory_virtual_repository.go
@@ -84,6 +84,20 @@ func resourceArtifactoryVirtualRepository() *schema.Resource {
 	return skeema
 }
 
+func resourceArtifactoryVirtualGenericRepository(pkt string) *schema.Resource {
+	constructor := func() interface{} {
+		return &VirtualRepositoryBaseParams{
+			PackageType: pkt,
+			Rclass:      "virtual",
+		}
+	}
+	unpack := func(data *schema.ResourceData) (interface{}, string, error) {
+		repo := unpackBaseVirtRepo(data, pkt)
+		return repo, repo.Id(), nil
+	}
+	return mkResourceSchema(baseVirtualRepoSchema, defaultPacker, unpack, constructor)
+}
+
 type DebianVirtualRepositoryParams struct {
 	VirtualRepositoryBaseParams
 	DebianTrivialLayout *bool `json:"debianTrivialLayout,omitempty"`
